feat: Simpler auth with axios, incl refreshing JWT
Simplifies the frontend/src/api dir significantly by removing manually
implemented http client and just using axios.

login and register functions are provided in auth.ts and connected to
the LoginPage.tsx and RegisterPage.tsx in the frontend.

An axios interceptor is provided in client.ts to automatically get a new
JWT when the current access token is expired/invalid.