feat: Protected/guest routes and logout button
It is now impossible to access dashboard/settings if logged out (user is
kicked to login page instead). Likewise, login/register pages are
inaccessible to authenticated users.

Also, the settings and user icon in the header render conditionally only
to authed users, and the user icon provides the option to log out.