feat(frontend): Improved toasts for deck creation
When a deck is successfully created, the toast provides a button to
start adding cards to that deck immediately.

If the deck fails due to a 409 Conflict error, the toast tells the user
a deck with that name already exists. Otherwise, a generic error
message.

Note that 409 Conflict is not yet implemented in the backend. Nothing
prevents decks with the same name from being created atm but this is
planned to change later.