feat(frontend): getDeck func and refactor in CreateCardPage
CreateCardPage was manually performing an axios get request to get the
name of a deck based on the deckId to show the deckname on the page.

Now there is a getDeck function in decks.ts which does this, and
CreateCardPage uses that instead.

Also includes another minor change whereby CreateCardPage shows an error
if you try to add cards to a deck that doesn't exist, e.g. by typing a
fake deckId into the URL bar.