feat(frontend): Edit decks and minor CreateCardPage change
The "Edit deck details" in each deck button's dropdown menu now invokes
a dialog to edit the deck namd and description. The dialog is a
refactored version of the CreateDeckDialog, now DeckEditDialog.

This also involved refactoring the info stored in each deck button. They
now store a Deck object as a prop instead of just deckName and deckId.

Also changed CreateCardPage to show a loading spinner where the deck
name would be ("Card will be saved to...") if the deck name is still
being fetched.